renamed interpolation/resampling header
